# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api app",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    },
                    "200": {
                        "description": "Successfully auth"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register by email, password",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "TEst"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials"
                    },
                    "200": {
                        "description": "Successfully auth"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                }
            }
        },
        "/api/carts": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Carts create",
                "description": "Create cart",
                "operationId": "cartCreate",
                "responses": {
                    "201": {
                        "description": "Cart created"
                    }
                }
            }
        },
        "/api/add-to-cart": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add to cart",
                "description": "Add item to cart",
                "operationId": "addItemToCart",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "uuid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "06424b27-3f90-4cb9-b32c-a7a0e23d834a"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cart exist. Item added"
                    },
                    "404": {
                        "description": "Cart not found Or Product not found"
                    }
                }
            }
        },
        "/api/cart-items": {
            "post": {
                "tags": [
                    "Cart items"
                ],
                "summary": "update cart item",
                "description": "update cart item",
                "operationId": "cartItemUpdate",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "uuid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "06424b27-3f90-4cb9-b32c-a7a0e23d834a"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "new_value": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "cart not found OR cart item not found"
                    },
                    "200": {
                        "description": "successfully updated"
                    }
                }
            }
        },
        "/api/cart-items/{id}": {
            "delete": {
                "tags": [
                    "Cart items"
                ],
                "summary": "delete cart item",
                "description": "delete cart item",
                "operationId": "cartItemDelete",
                "requestBody": {
                    "description": "Pass user credentials"
                },
                "responses": {
                    "404": {
                        "description": "cart item not found"
                    },
                    "200": {
                        "description": "successfully delete"
                    },
                    "400": {
                        "description": "delete error"
                    }
                }
            }
        }
    }
}